optimized counting of contingency tables
Most steps of the Monte Carlo random walk are invariant, so instead of sampling and "performing" all of them, we sample from a geometric distribution the number of invariant steps that are expected, and then perform a step that is guaranteed to not be invariant. This saves a lot of iterations on very sparse matrices.